Move small formFactor tests to a different file
