--- force-app/test/jest-mocks/@salesforce/client/formFactor.js
@@ -1,3 +1,0 @@
-const FORM_FACTOR = 'Large';
-
-export default FORM_FACTOR;
