+++ jest.config.js
@@ -6,8 +6,6 @@ setupFilesAfterEnv.push('<rootDir>/jest-sa11y-setup.js');
 module.exports = {
     ...jestConfig,
     moduleNameMapper: {
-        '^@salesforce/client/formFactor$':
-            '<rootDir>/force-app/test/jest-mocks/@salesforce/client/formFactor',
         '^lightning/navigation$':
             '<rootDir>/force-app/test/jest-mocks/lightning/navigation',
         '^lightning/platformShowToastEvent$':
